Lots of refactoring
