+++ app/src/main/java/projects/medicationtracker/AddMedication.java
@@ -1,6 +1,6 @@
 package projects.medicationtracker;
 
-import static projects.medicationtracker.TimeFormatting.stringToLocalDateTime;
+import static projects.medicationtracker.Helpers.TimeFormatting.stringToLocalDateTime;
 
 import android.annotation.SuppressLint;
 import android.icu.text.SimpleDateFormat;
@@ -36,6 +36,15 @@ import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Locale;
 import java.util.Objects;
+
+import projects.medicationtracker.Fragments.SelectDateFragment;
+import projects.medicationtracker.Fragments.TimePickerFragment;
+import projects.medicationtracker.Helpers.DBHelper;
+import projects.medicationtracker.Helpers.InputValidation;
+import projects.medicationtracker.Helpers.NotificationHelper;
+import projects.medicationtracker.Helpers.TextViewUtils;
+import projects.medicationtracker.Helpers.TimeFormatting;
+import projects.medicationtracker.SimpleClasses.Medication;
 
 public class AddMedication extends AppCompatActivity
 {
