+++ app/src/main/java/projects/medicationtracker/EditMedication.java
@@ -26,6 +26,16 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.Objects;
 
+import projects.medicationtracker.Fragments.ConfirmMedicationDeleteFragment;
+import projects.medicationtracker.Fragments.SelectDateFragment;
+import projects.medicationtracker.Fragments.TimePickerFragment;
+import projects.medicationtracker.Helpers.DBHelper;
+import projects.medicationtracker.Helpers.InputValidation;
+import projects.medicationtracker.Helpers.NotificationHelper;
+import projects.medicationtracker.Helpers.TextViewUtils;
+import projects.medicationtracker.Helpers.TimeFormatting;
+import projects.medicationtracker.SimpleClasses.Medication;
+
 public class EditMedication extends AppCompatActivity
 {
     final DBHelper db = new DBHelper(this);
