+++ app/src/main/java/projects/medicationtracker/Fragments/ConfirmMedicationDeleteFragment.java
@@ -1,4 +1,4 @@
-package projects.medicationtracker;
+package projects.medicationtracker.Fragments;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
@@ -7,6 +7,11 @@ import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.DialogFragment;
+
+import projects.medicationtracker.Helpers.DBHelper;
+import projects.medicationtracker.Helpers.NotificationHelper;
+import projects.medicationtracker.SimpleClasses.Medication;
+import projects.medicationtracker.MyMedications;
 
 public class ConfirmMedicationDeleteFragment extends DialogFragment
 {
@@ -18,7 +23,7 @@ public class ConfirmMedicationDeleteFragment extends DialogFragment
      * @param database  The DBHelper connected to the database with the medication.
      * @param medicationToDelete The medication to delete from the database.
      **************************************************************************/
-    ConfirmMedicationDeleteFragment(DBHelper database, Medication medicationToDelete)
+    public ConfirmMedicationDeleteFragment(DBHelper database, Medication medicationToDelete)
     {
         db = database;
         medication = medicationToDelete;
