+++ app/src/main/java/projects/medicationtracker/Fragments/EditNoteFragment.java
@@ -1,4 +1,4 @@
-package projects.medicationtracker;
+package projects.medicationtracker.Fragments;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
@@ -10,6 +10,11 @@ import android.widget.EditText;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.DialogFragment;
+
+import projects.medicationtracker.Helpers.DBHelper;
+import projects.medicationtracker.MedicationNotes;
+import projects.medicationtracker.SimpleClasses.Note;
+import projects.medicationtracker.R;
 
 public class EditNoteFragment extends DialogFragment
 {
