+++ app/src/main/java/projects/medicationtracker/Helpers/InputValidation.java
@@ -1,4 +1,4 @@
-package projects.medicationtracker;
+package projects.medicationtracker.Helpers;
 
 import android.widget.EditText;
 import android.widget.RadioButton;
