+++ app/src/main/java/projects/medicationtracker/MainActivity.java
@@ -25,6 +25,13 @@ import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.Objects;
 
+import projects.medicationtracker.Helpers.DBHelper;
+import projects.medicationtracker.Helpers.NotificationHelper;
+import projects.medicationtracker.Helpers.TextViewUtils;
+import projects.medicationtracker.Helpers.TimeFormatting;
+import projects.medicationtracker.SimpleClasses.Medication;
+import projects.medicationtracker.Views.StandardCardView;
+
 public class MainActivity extends AppCompatActivity
 {
     private final DBHelper db = new DBHelper(this);
