--- app/src/main/java/projects/medicationtracker/ScheduleSlidePagerActivity.java
@@ -1,8 +1,0 @@
-package projects.medicationtracker;
-
-import androidx.fragment.app.FragmentActivity;
-
-public class ScheduleSlidePagerActivity extends FragmentActivity
-{
-
-}
