+++ app/src/main/java/projects/medicationtracker/SimpleClasses/Medication.java
@@ -1,4 +1,4 @@
-package projects.medicationtracker;
+package projects.medicationtracker.SimpleClasses;
 
 import androidx.annotation.NonNull;
