+++ app/src/main/java/projects/medicationtracker/SimpleClasses/Note.java
@@ -1,4 +1,4 @@
-package projects.medicationtracker;
+package projects.medicationtracker.SimpleClasses;
 
 import java.time.LocalDateTime;
