+++ app/src/main/res/layout/activity_edit_medication.xml
@@ -22,7 +22,11 @@
             <androidx.cardview.widget.CardView
                 android:id="@+id/editMedCard"
                 android:layout_width="match_parent"
-                android:layout_height="match_parent">
+                android:layout_height="match_parent"
+                android:layout_marginLeft="5dp"
+                android:layout_marginTop="5dp"
+                android:layout_marginRight="5dp"
+                app:cardCornerRadius="5dp">
 
                 <TableLayout
                     android:id="@+id/medDetailsTable"
