+++ app/src/main/res/layout/fragment_add_note.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".AddNoteFragment">
+    tools:context=".Fragments.AddNoteFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
