+++ app/src/main/res/layout/fragment_edit_note.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditNoteFragment">
+    tools:context=".Fragments.EditNoteFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
